Refactor createpost.component.ts -> blogpost.create.component.ts.
