+++ frontend/src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { LoginComponent } from './login/login.component';
 import { BlogpostComponent } from './blogpost/blogpost.component';
-import { CreatePostComponent } from './blogpost/createpost.component';
+import { CreatePostComponent } from './blogpost/blogpost.create.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent },
